Fix to support mainline kexec dtb
Mainline kernel passes device tree blob in r2, not in ATAGs. Also the
kexec_mach_type should be 0xffffffff for the mainline kernel for device
tree based booting.

We want to use mainline kernel as is with dtb file and initramfs,
so let's do minimal fixes to support passing dtb in r2. And let's
pass the generic DT machine as the machine type. This allows us to
use the current mainline kexec-tools too.

So basically this patch is just a backport of the current mainline
kernel kexec code that updates the kexec_mach_type based on what kind
of device tree is detected. This means we can kexec boot both mainline
kernels with standard dtb, and legacy kernels using custom Motorola
little-endian dtb.

When attempting to kexec boot legacy kernels, note that probably the
stock kernel won't work for kexec properly. And note that the Lineage
OS initrd seems to depend on SafeStrap config being there for loopback
devices, and needs to use the custom legacy kexec binary with atags.

+++ kernel/time.c
@@ -246,7 +246,7 @@ inline unsigned int jiffies_to_msecs(const unsigned long j)
 	return (j + (HZ / MSEC_PER_SEC) - 1)/(HZ / MSEC_PER_SEC);
 #else
 # if BITS_PER_LONG == 32
-	return (HZ_TO_MSEC_MUL32 * j) >> HZ_TO_MSEC_SHR32;
+	return ((u64)HZ_TO_MSEC_MUL32 * j) >> HZ_TO_MSEC_SHR32;
 # else
 	return (j * HZ_TO_MSEC_NUM) / HZ_TO_MSEC_DEN;
 # endif
@@ -262,7 +262,7 @@ inline unsigned int jiffies_to_usecs(const unsigned long j)
 	return (j + (HZ / USEC_PER_SEC) - 1)/(HZ / USEC_PER_SEC);
 #else
 # if BITS_PER_LONG == 32
-	return (HZ_TO_USEC_MUL32 * j) >> HZ_TO_USEC_SHR32;
+	return ((u64)HZ_TO_USEC_MUL32 * j) >> HZ_TO_USEC_SHR32;
 # else
 	return (j * HZ_TO_USEC_NUM) / HZ_TO_USEC_DEN;
 # endif
@@ -467,7 +467,7 @@ unsigned long msecs_to_jiffies(const unsigned int m)
 	if (HZ > MSEC_PER_SEC && m > jiffies_to_msecs(MAX_JIFFY_OFFSET))
 		return MAX_JIFFY_OFFSET;
 
-	return (MSEC_TO_HZ_MUL32 * m + MSEC_TO_HZ_ADJ32)
+	return ((u64)MSEC_TO_HZ_MUL32 * m + MSEC_TO_HZ_ADJ32)
 		>> MSEC_TO_HZ_SHR32;
 #endif
 }
@@ -482,7 +482,7 @@ unsigned long usecs_to_jiffies(const unsigned int u)
 #elif HZ > USEC_PER_SEC && !(HZ % USEC_PER_SEC)
 	return u * (HZ / USEC_PER_SEC);
 #else
-	return (USEC_TO_HZ_MUL32 * u + USEC_TO_HZ_ADJ32)
+	return ((u64)USEC_TO_HZ_MUL32 * u + USEC_TO_HZ_ADJ32)
 		>> USEC_TO_HZ_SHR32;
 #endif
 }
